requirements.txt
Django==4.2.2
djangorestframework==3.14.0
django-cors-headers==4.0.0
django_filter==23.1